Refatored Main routes
